Split the reference state out into BackgroundPotentialEnergyEquation
The reference state and the background potential energy now live in their own
module. `reference_height`, `reference_buoyancy`, `BackgroundPotentialEnergy`
and the four reference-height methods are defined there and re-exported by
`AvailablePotentialEnergyEquation`, so either module can be used on its own.
`AvailablePotentialEnergyEquation` keeps only the local Eₐ kernel and its
constructors, and imports the rest.

Ψ = ∫b✶dz̃ moves with the reference state rather than with Eₐ: it is a property
of the reference profile, and computing it anywhere else would mean sorting a
second time. Renamed to `reference_potential` accordingly, since "ape" no
longer describes where it lives.

The docs page splits the same way. The reference-state discussion, the four
methods, and the cost benchmark move to a new Background potential energy
equation page alongside the `BackgroundPotentialEnergy` and reference-height
docstrings; the Available page keeps the Eₐ definition and points at the other
for z✶. Cross-page `@ref`s resolve in both directions, checked with
cross_references strict.

Also carries three earlier plotting changes to the lock-release example that
had not been committed: z runs 0 to H rather than -H/2 to H/2, the reference
profiles are a 2x2 grid, and the heatmaps use DataAspect.
